refactored parser to `http` module and enhanced tests
