+++ src/parser/mod.rs
@@ -1,2 +1,1 @@
-mod de;
-pub mod parser;
+pub mod de;
